商品購入機能実装

+++ db/migrate/20200925095500_create_addresses.rb
@@ -7,7 +7,7 @@ class CreateAddresses < ActiveRecord::Migration[6.0]
       t.string      :city,            null:false
       t.string      :address,         null:false
       t.string      :building_name
-      t.integer     :phone_number,    null:false
+      t.bigint      :phone_number,    null:false
       
       t.references  :order,           null:false,  foreign_key: true
